Module type Mbt.MEASURE

Type of balanced trees with a measure

Measure is a value from a monoid that is computed from payloads and accumulated along branches until reaching the root of the tree.

type +'a measurable

type measure

val empty : measure

val cat : measure -> 'a measurable -> measure -> measure