grenier index

Library grenier.balmap

The entry point of this library is the module: Balmap.

Library grenier.baltree

This library exposes the following toplevel modules:

Bt1

Bt2

Mbt

Library grenier.binder_introducer

The entry point of this library is the module: Binder_introducer.

Library grenier.binpacking

The entry point of this library is the module: Maxrects.

Library grenier.congre

The entry point of this library is the module: Congre.

Library grenier.dbseq

The entry point of this library is the module: Dbseq.

Library grenier.doubledouble

The entry point of this library is the module: Doubledouble.

Library grenier.fastdom

The entry point of this library is the module: Fastdom.

Library grenier.hll

This library exposes the following toplevel modules:

Hll An implementation of HyperLogLog probabilistic cardinality estimator.

Hll_consts

Library grenier.jmphash

The entry point of this library is the module: Jmphash.

Library grenier.orderme

This library exposes the following toplevel modules:

Order_indir

Order_interval

Order_list

Order_managed

Order_managed_indir

Order_managed_interval

Library grenier.state_elimination

The entry point of this library is the module: State_elimination.

Library grenier.strong

The entry point of this library is the module: Strong.

Library grenier.trope

The entry point of this library is the module: Trope.

Library grenier.valmari

This library exposes the following toplevel modules:

Partition

Valmari Valmari is an automata minimization algorithm, described in "Fast brief practical DFA minimization" https://dl.acm.org/doi/10.1016/j.ipl.2011.12.004